+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Version 2.x
 
+# 2.2.4
+
+* Add class union types in reflection parser in order to support serialization without using a `UnionDiscrimnator` attribute.
+
 # 2.2.3
 
 * Make sure the `DoctrineMetadataParser` can handle metadata from version `2.x` and `3.x` of the package `doctrine/orm`